requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dicom2nifti==2.5.1
docopt==0.6.2
exceptiongroup==1.2.2
fastapi==0.115.6
fonttools==4.55.3
h11==0.14.0
idna==3.10
importlib-resources==6.5.2
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.4
nibabel==5.3.2
numpy==1.26.4
packaging==24.2
pillow==11.1.0
pydantic==2.10.5
pydantic-core==2.27.2
pydicom==2.4.4
pykwalify==1.8.0
pyparsing==3.2.1
pyradiomics==3.1.0
python-dateutil==2.9.0.post0
python-gdcm==3.0.24.1
python-multipart==0.0.20
pywavelets==1.6.0
requests==2.32.3
ruamel-yaml==0.18.10
ruamel-yaml-clib==0.2.12
scikit-learn==1.6.1
scipy==1.13.1
simpleitk==2.4.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
threadpoolctl==3.5.0
typing-extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
zipp==3.21.0
